Split config functions into own file
